+++ lib/dnsimple/struct/domain.rb
@@ -4,13 +4,6 @@ module Dnsimple
   module Struct
 
     class Domain < Base
-
-      def initialize(attributes = {})
-        attributes.delete("expires_on")
-        super
-        @expires_on = Date.parse(expires_at).to_s if expires_at
-      end
-
       # @return [Integer] The domain ID in DNSimple.
       attr_accessor :id
 
@@ -43,19 +36,6 @@ module Dnsimple
 
       # @return [String] When the domain was last updated in DNSimple.
       attr_accessor :updated_at
-
-      # @deprecated Please use #expires_at instead.
-      # @return [String] The date the domain will expire.
-      def expires_on
-        warn "[DEPRECATION] Domain#expires_on is deprecated. Please use `expires_at` instead."
-        @expires_on
-      end
-
-      def expires_on=(expiration_date)
-        warn "[DEPRECATION] Domain#expires_on= is deprecated. Please use `expires_at=` instead."
-        @expires_on = expiration_date
-      end
-
     end
   end
 end