+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 - NEW: Added `registrar.get_domain_prices` to retrieve whether a domain is premium and the prices to register, transfer, and renew. (dnsimple/dnsimple-ruby#230)
 - REMOVED: `domain.reset_domain_token` endpoint no longer exists and the client method is removed.
   (dnsimple/dnsimple-ruby#231)
+- REMOVED: The deprecated `Domain.expires_on` is removed. (dnsimple/dnsimple-ruby#232)
+- REMOVED: The deprecated `Certificate.expires_on` is removed. (dnsimple/dnsimple-ruby#232)
 
 ## 5.2.0